+++ gilded-rose-web/src/main/java/com/gildedrose/web/dto/ItemDTO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class ItemDTO {
 
-    public int id;
+    public long id;
     public String name;
     public int sellIn;
     public int quality;
@@ -12,7 +12,7 @@ public class ItemDTO {
 
     public ItemDTO() {
     }
-    public ItemDTO(int id, String name, int sellIn, int quality) {
+    public ItemDTO(long id, String name, int sellIn, int quality) {
         this.id = id;
         this.name = name;
         this.sellIn = sellIn;
